fixed suggested comments
