+++ packages/components/acter/landing-page/members-section/connection-state.tsx
@@ -14,7 +14,6 @@ export interface ConnectionStateProps
 }
 
 export const ConnectionState: FC<ConnectionStateProps> = ({
-  acter,
   connection,
   canEdit,
   onSubmit,
@@ -28,7 +27,6 @@ export const ConnectionState: FC<ConnectionStateProps> = ({
   if (canEdit && isEditing) {
     return (
       <ConnectionStateEditor
-        acter={acter}
         connection={connection}
         onCancel={() => setIsEditing(false)}
         onSubmit={onSubmit}
